+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 **v0.2.1 - 05-07-2026**
 ---
 
+- Added thread and message focus mode
 - Added French translation
 - Fixed untranslated drop down menu items
 - Fixed iCloud Empty Trash [#136](https://github.com/hkdb/aerion/issues/136)